// src/app/components/model/state-interface.ts
export interface StateInterface {
  name: string;
  code: string;
  stateCode: string;
  selected: boolean;
  totalPopulation: number;
  medianIncome: number;
  peopleInPoverty: number;
  unemployed: number;
  employed: number;
  medianHomeValue: number;
  medianRent: number;
  universityDegree: number;
  noHealthInsurance: number;
}